moved HhDrawer to separate directory
