+++ lib/app/events/events_view.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:hnh/app/abstract/view.dart';
 import 'package:hnh/app/components/event_card.dart';
-import 'package:hnh/app/components/hhDrawer.dart';
 import 'package:hnh/app/events/events_controller.dart';
 import 'package:modal_progress_hud/modal_progress_hud.dart';
 import 'package:hnh/app/utils/constants.dart';
