+++ lib/app/map/map_view.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:hnh/app/components/hhDrawer.dart';
 import 'package:hnh/app/map/map_controller.dart';
 import 'package:hnh/app/abstract/view.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
